Removed Note About Enforcement
This is the only rule that mentions enforcement. While those of us involved know that all of these guidelines will be enforced, specifically mentioning enforcement on only one guideline has the potential to create an expectation that it is the only enforced guideline and therefore the only one worth respecting.