Clarify ‘extortion’
Sadly that has to remain as it’s not a hypothetical situation. Added in pressuring and extortion, as both of those should be considered as well. I wish the tl;dr of ‘don’t be mean!’ was enough, but we’ve had enough situations where this has to be spelled out.

Props @dmccan

Fixes #27