Clarify GPL and Responsibility
G1 - Change GPL wording to be less ambigious — Props @rarst

G2 - Change title and phrasing to explicitly state that developers accept sole responsibility for what they upload. It’s their job to comply to the guidelines and ours to enforce it. Should be obvious. Wish it was. It’s not. — Props @otto42