Tightening Language
One thing here, it's better to not specific a number for a guideline if 
we don't have to, so if we renumber them, we don't go insane fixing it.

Intro: 'are expected to' -> MUST. You MUST follow the guidelines.
01. "or later" added becuase that's literally what it is. Removing 
duplicate sentance.
02. Remove 'sole' as that is misundertood in some languages to mean it's 
their ONLY responsibility, instead of 'its your own responsibility.' 
Correcting bad grammar.
03. Clarify that you can dev where you want, but only what we host is 
what user's get.
05. Remove parenthetical in favor of a complete sentance.
06. Remove unneeded reitteration. We already said 'things that aren't a 
service' so leave the dead horse alone.
07. Emphasis on PROHIBITED
08. Bad grammar and run on sentances.
09. Formatting and grammar. Clarifying compliance.
10. Removing extranous words.
11. Prioritizing the WHY of the gudeline before devling into the 
details. It may be better to list what kinds of ads are not allowed, but 
that would be a complicated list to curate. Attempting to explain with 
an example ...
14. Breaking apart sentances and paragraphs for readability. Removing 
pejorative for 'trash' commits.
17. Grammar